updates for robustness
Comb through the entire pulse audio device list for possible sinks, as
in Ubuntu the same audio device seems to jump around after kernel
updates, making the mute / unmute stop working.

Also, don't do the xchat calls if xchat is not running. This prevents
us from spawning useless xchat windows in the event that we've turned
off xchat because we're on paternity leave, but are still using the
computer.